一些细节优化
注册界面可以转到登陆；自动分割后可选择会首页或继续自动分割下一张；未登陆不能进行分割操作

+++ flask_web_with_sql/upload.py
@@ -38,26 +38,34 @@ def allowed_file(filename):
 app.send_file_max_age_default = datetime.timedelta(seconds=1)
 basepath = os.path.dirname(__file__)
 
+logined = False
 # 用户登录
 @app.route('/login', methods=['GET', 'POST'])
 def login():
     if request.method == 'POST':
-        username_login = request.form['username']
-        password_login = request.form['password']
+        post_value_exists = False
+        for check_post_value in request.form:
+            post_value_exists = True
         
-        user_login = class_user(username_login)
-        user_login.connect_to_database(host,username,password,database_name)
-        user_login.user_check(user_table_name)
-        user_login.disconnect_database()
-        
-        if user_login.exist == False:
-            return jsonify({'status': '-1', 'errmsg': 'user does not exist!'})
-        elif password_login == user_login.Password:
-            global user_temp
-            user_temp = user_login.User
-            return redirect(url_for('choose'))  # url_for后面加的是函数名
-        else :
-            return jsonify({'status': '-1', 'errmsg': 'password is wrong!'})
+        if post_value_exists == True:
+            username_login = request.form['username']
+            password_login = request.form['password']
+            
+            user_login = class_user(username_login)
+            user_login.connect_to_database(host,username,password,database_name)
+            user_login.user_check(user_table_name)
+            user_login.disconnect_database()
+            
+            if user_login.exist == False:
+                return jsonify({'status': '-1', 'errmsg': 'user does not exist!'})
+            elif password_login == user_login.Password:
+                global user_temp
+                user_temp = user_login.User
+                global logined
+                logined = True
+                return redirect(url_for('choose'))  # url_for后面加的是函数名
+            else :
+                return jsonify({'status': '-1', 'errmsg': 'password is wrong!'})
 
     return render_template('login.html')
 
@@ -91,6 +99,9 @@ def register():
 # 选择分割方式，radio实现
 @app.route('/draw', methods=['get','post'])  # form表单中的action对应的是 网址！！不是函数名
 def draw():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     if request.method == 'POST':
         path = basepath + "/static/manual_photos/manual_picture_window_changed.jpg"
         file_path = path
@@ -120,6 +131,9 @@ def choose():
 # 选择分割方式，radio实现
 @app.route('/windowChoose', methods=['get','post'])  # form表单中的action对应的是 网址！！不是函数名
 def windowChoose():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     if request.method == 'POST':
         centerSet = request.form['center']
         widthSet = request.form['width']
@@ -135,25 +149,33 @@ def windowChoose():
 # 自动分割处理图片上传
 @app.route('/auto_segment', methods=['POST', 'GET'])   # 添加路由
 def auto_segment():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     if request.method == 'POST':
-        f = request.files['file']
-        path = basepath + "/static/auto_photos/"
-        file_path = path + f.filename  # 图片路径和名称
-        print(file_path)
-        f.save(file_path)  # 保存图片
-        # 到本地文件读取图片并且显示在网页上
-        (shotname, extension) = os.path.splitext(f.filename);
-        if (extension == '.dcm'):
-            dcm = pydicom.read_file(file_path)
-            img = dcm.pixel_array
-            img = np.float32(img)
-        else:
-            img = cv2.imread(file_path)
-        # 根据图片名字进行存储，但显示有问题
-        # cv2.imwrite(os.path.join(file_path), img)   # 保存图片，第一个参数是路径加图像名，第二个是图像矩阵
-        # 将图片名字都更改为test.jpg
-        cv2.imwrite(os.path.join(path, 'auto_picture.jpg'), img)
-        return render_template('auto_segment_upload.html')
+        post_file_exists = False
+        for check_post_file in request.files:
+            post_file_exists = True
+            
+        if post_file_exists == True:
+            f = request.files['file']
+            path = basepath + "/static/auto_photos/"
+            file_path = path + f.filename  # 图片路径和名称
+            print(file_path)
+            f.save(file_path)  # 保存图片
+            # 到本地文件读取图片并且显示在网页上
+            (shotname, extension) = os.path.splitext(f.filename);
+            if (extension == '.dcm'):
+                dcm = pydicom.read_file(file_path)
+                img = dcm.pixel_array
+                img = np.float32(img)
+            else:
+                img = cv2.imread(file_path)
+            # 根据图片名字进行存储，但显示有问题
+            # cv2.imwrite(os.path.join(file_path), img)   # 保存图片，第一个参数是路径加图像名，第二个是图像矩阵
+            # 将图片名字都更改为test.jpg
+            cv2.imwrite(os.path.join(path, 'auto_picture.jpg'), img)
+            return render_template('auto_segment_upload.html')
     return render_template('auto_segment.html')
 
 # 自动分割后的图片显示
@@ -162,12 +184,18 @@ def auto_segment():
 # 选择分割类型(腹部 or  颅脑  or  左心室)
 @app.route('/segmentation_type', methods=['get','post'])  # form表单中的action对应的是 网址！！不是函数名
 def segmentation_type():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     segment = request.values.get('segment_type')
     if segment == 'AAT':
          return redirect(url_for('auto_belly_segment'))  # url_for后面加的是函数名
 
 @app.route('/auto_belly_segment')
 def auto_belly_segment():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     mPredict.LoadPic()
     mPredict.PredictPic()
     mPredict.SavePic()
@@ -176,6 +204,9 @@ def auto_belly_segment():
 # 手动分割处理
 @app.route('/manual_segment', methods=['POST', 'GET'])   # 添加路由
 def manual_segment():
+    global logined
+    if logined == False:
+        return redirect(url_for('login'))
     if request.method == 'POST':
         f = request.files['file']
         path = basepath + "/static/manual_photos/"
